Extract SchemaDoc from DocExplorer
