+++ src/components/DocExplorer.js
@@ -18,6 +18,7 @@ import {
 } from 'graphql';
 
 import MarkdownContent from './DocExplorer/MarkdownContent';
+import SchemaDoc from './DocExplorer/SchemaDoc';
 import SearchBox from './DocExplorer/SearchBox';
 import SearchResults from './DocExplorer/SearchResults';
 import TypeLink from './DocExplorer/TypeLink';
@@ -189,67 +190,6 @@ export class DocExplorer extends React.Component {
       name: 'Search Results',
       searchValue: value
     });
-  }
-}
-
-// Render the top level Schema
-class SchemaDoc extends React.Component {
-
-  static propTypes = {
-    schema: PropTypes.object,
-    onClickType: PropTypes.func,
-  }
-
-  shouldComponentUpdate(nextProps) {
-    return this.props.schema !== nextProps.schema;
-  }
-
-  render() {
-    const schema = this.props.schema;
-    const queryType = schema.getQueryType();
-    const mutationType = schema.getMutationType && schema.getMutationType();
-    const subscriptionType =
-      schema.getSubscriptionType && schema.getSubscriptionType();
-
-    return (
-      <div>
-        <MarkdownContent
-          className="doc-type-description"
-          markdown={
-            'A GraphQL schema provides a root type for each kind of operation.'
-          }
-        />
-        <div className="doc-category">
-          <div className="doc-category-title">
-            {'root types'}
-          </div>
-          <div className="doc-category-item">
-            <span className="keyword">{'query'}</span>
-            {': '}
-            <TypeLink type={queryType} onClick={this.props.onClickType} />
-          </div>
-          {
-            mutationType &&
-            <div className="doc-category-item">
-              <span className="keyword">{'mutation'}</span>
-              {': '}
-              <TypeLink type={mutationType} onClick={this.props.onClickType} />
-            </div>
-          }
-          {
-            subscriptionType &&
-            <div className="doc-category-item">
-              <span className="keyword">{'subscription'}</span>
-              {': '}
-              <TypeLink
-                type={subscriptionType}
-                onClick={this.props.onClickType}
-              />
-            </div>
-          }
-        </div>
-      </div>
-    );
   }
 }
